refactor: user 엔티티 경로 변경으로 인한 import 수정
