+++ src/main/java/com/example/popin/domain/space/entity/Space.java
@@ -1,6 +1,6 @@
 package com.example.popin.domain.space.entity;
 
-import com.example.popin.domain.user.User;
+import com.example.popin.domain.user.entity.User;
 import com.example.popin.global.common.BaseEntity;
 import javax.persistence.*;
 import lombok.AccessLevel;
