+++ src/main/java/com/example/popin/domain/space/repository/SpaceRepository.java
@@ -1,7 +1,7 @@
 package com.example.popin.domain.space.repository;
 
 import com.example.popin.domain.space.entity.Space;
-import com.example.popin.domain.user.User;
+import com.example.popin.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
